imporve audio capture

+++ src/ndp_flash.c
@@ -6,9 +6,11 @@
  *
  *****************************************************************************/
 #include <stdio.h>
+#include <string.h>
 #include "syntiant_platform.h"
 #include "ndp_flash.h"
 #include "fat_load.h"
+#include <string.h>
 
 #define SYNTIANT_NDP_ERROR_NONE  0
 #define DUMMY_0_BYTES			0
@@ -381,6 +383,33 @@ static int ndp_flash_program_firmware(uint32_t address, char * file_name)
     return ret;
 }
 
+/**
+ * mode: 4 bytes
+ * buttonswitch: length: 4 bytes, data_string: length
+*/
+#define FLASH_MAX_ADDRESS       0x100000
+#define FLASH_INFO_ADDRESS    (FLASH_MAX_ADDRESS-FLASH_SPLIT_SIZE)  
+static int ndp_flash_program_infos(int mode_val, char *button_val, 
+        uint32_t button_val_len)
+{
+	int ret = -1;
+    uint8_t burn_data[FLASH_SPLIT_SIZE];
+	uint32_t burn_addr = FLASH_INFO_ADDRESS;
+    uint32_t burn_len = 0;
+
+	printf("FLASH programming mode: %d, button_swith: %s \n",mode_val, button_val);
+
+    *(uint32_t*)&burn_data[burn_len] = mode_val;
+    burn_len += sizeof(uint32_t);
+    *(uint32_t*)&burn_data[burn_len] = strlen(button_val);
+    burn_len += sizeof(uint32_t);
+    strncpy((char*)&burn_data[burn_len], button_val, button_val_len);
+    burn_len += strlen(button_val);
+
+	ret = ndp_flash_write_block(burn_addr, burn_data, burn_len);
+    return ret;
+}
+
 int ndp_flash_program_all_fw(void)
 {
     int ret;
@@ -399,10 +428,33 @@ int ndp_flash_program_all_fw(void)
 
 	//flash boot file
 	ret = ndp_flash_program_firmware(0x00, flash_file_name);
+    if (ret) return ret;
 
 	//remove file
 	printf("remove %s \n", flash_file_name);
 	remove_file(flash_file_name);
+
+    ret = ndp_flash_program_infos(mode_index, button_switch, strlen(button_switch));
+
+    return ret;
+}
+
+int ndp_flash_read_infos(int *mode_val, char *button_val)
+{
+    int ret = 0;
+    uint8_t read_data[FLASH_SPLIT_SIZE];
+    uint32_t read_offset = 0;
+    uint32_t button_val_len = 0;
+
+    ret = ndp_flash_read_block(FLASH_INFO_ADDRESS, read_data, FLASH_SPLIT_SIZE);
+    if (ret) return ret;
+
+    *mode_val = *(uint32_t*)&read_data[read_offset];
+    read_offset += sizeof(uint32_t);
+
+    button_val_len = *(uint32_t*)&read_data[read_offset];
+    read_offset += sizeof(uint32_t);
+    strncpy(button_val, (char*)&read_data[read_offset], button_val_len);
 
     return ret;
 }
